refacto: move loaders with little to no dependency to own directory
